# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  routes.ErrorResponse:
    properties:
      message:
        example: An error occurred.
        type: string
    type: object
  routes.LoginRequest:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  routes.LoginResponse:
    properties:
      message:
        example: Login successful!
        type: string
      token:
        example: token123
        type: string
    type: object
  routes.SignUpRequest:
    properties:
      email:
        example: test@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  routes.SignUpResponse:
    properties:
      message:
        example: User created.
        type: string
      user:
        example: test@example.com
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample API for a travel advisor application built with Golang.
  license:
    name: MIT License
    url: https://opensource.org/license/mit/
  title: Golang Travel Advisor API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token.
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/routes.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Login successful."  example({"message": "Login successful!",
            "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."})'
          schema:
            $ref: '#/definitions/routes.LoginResponse'
        "400":
          description: Could not parse request data.
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Wrong user credentials.
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: User login
      tags:
      - users
  /signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account.
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/routes.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'User created."  example({"message": "User created.", "user":
            "user@example.com"})'
          schema:
            $ref: '#/definitions/routes.SignUpResponse'
        "400":
          description: Could not parse request data or user already exists.
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Could not create user. Try again later.
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Register a new user
      tags:
      - users
securityDefinitions:
  Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"